commit 14.08.22 allure report (screen, test description, description, link), retry,test listener, builder (created User class), property reader

+++ src/test/java/com/prestashop/testing/utils/PropertyReader.java
@@ -17,7 +17,7 @@ public class PropertyReader {
         }
     }
 
-    public static int getTimeout(){
-        return Integer.parseInt(properties.getProperty("timeout"));
+    public static long getTimeout(){
+        return Long.parseLong(properties.getProperty("timeout"));
     }
 }
